+++ CHANGELOG.md
@@ -22,4 +22,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Add template for PRs  - [#53](https://github.com/ios-course/ironfoudation-team-project/pull/53)
 - Add template for issues  - [#54](https://github.com/ios-course/ironfoudation-team-project/pull/54)
 - Add missed information to `CHANGELOG.md`  - [#56](https://github.com/ios-course/ironfoudation-team-project/pull/56)
-- Add the `LocalStorage` class to manage persistent data storage - [58](https://github.com/ios-course/ironfoudation-team-project/pull/58)
+- Add the `KeyValueStorage` class to manage persistent data storage - [58](https://github.com/ios-course/ironfoudation-team-project/pull/58)